Avoid calling time.elapsed() twice in abstract animation

+++ src/corelib/animation/qabstractanimation_p.h
@@ -172,7 +172,7 @@ private:
 
     ElapsedTimer time;
 
-    int lastTick;
+    qint64 lastTick;
     int timingInterval;
     int currentAnimationIdx;
     bool consistentTiming;
